[#4] refactor: 데이터소스 분리
